Fix: Calculate correct length of adaption field
Calculates the length of adaption field , including
the byte for the adaption_field_length.

In the adaptation field, the first byte is the
adaptation_field_length. This field is the number of bytes
immediately following this byte. The result is that
the entire adaption field consumed N + 1 bytes. N being
the adaption_field_length.

Previously, the 1 byte was added ONLY if the
adaption_field_length was greater than 0. However, the adaption
field can be present and adaption_field_length can be 0. In this
case, the length of adaption field is 1.

Reference: https://en.wikipedia.org/wiki/MPEG_transport_stream